pyproject.toml
[project]
name = "stamp-testtask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "onnxruntime>=1.22.0",
    "opencv-python>=4.11.0.86",
    "pyyaml>=6.0.2",
    "ultralytics>=8.3.160",
]
